requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
aiodns==3.0.0
    # via discord-py
aiohttp==3.8.1
    # via
    #   discord-py
    #   tweepy
aiomysql==0.1.1
    # via -r requirements.in
aiosignal==1.2.0
    # via aiohttp
arrow==1.2.2
    # via -r requirements.in
astunparse==1.6.3
    # via import-expression
async-lru==1.0.3
    # via tweepy
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via aiohttp
black==22.6.0
    # via -r requirements.in
braceexpand==0.1.7
    # via jishaku
brotli==1.0.9
    # via discord-py
cchardet==2.1.7
    # via discord-py
certifi==2022.6.15
    # via requests
cffi==1.15.1
    # via pycares
charset-normalizer==2.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   black
    #   jishaku
discord-ext-menus @ git+https://github.com/joinemm/discord-ext-menus.git
    # via -r requirements.in
discord-ext-menus-views @ git+https://github.com/oliver-ni/discord-ext-menus-views
    # via -r requirements.in
discord.py @ git+https://github.com/Rapptz/discord.py.git
    # via
    #   -r requirements.in
    #   discord-ext-menus
flake8==4.0.1
    # via -r requirements.in
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.3
    # via
    #   requests
    #   yarl
import-expression==1.1.4
    # via jishaku
jishaku==2.5.0
    # via -r requirements.in
mccabe==0.6.1
    # via flake8
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via black
oauthlib==3.2.0
    # via
    #   requests-oauthlib
    #   tweepy
orjson==3.7.7
    # via discord-py
pathspec==0.9.0
    # via black
platformdirs==2.5.2
    # via black
psutil==5.9.1
    # via -r requirements.in
pycares==4.2.1
    # via aiodns
pycodestyle==2.8.0
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==2.4.0
    # via flake8
pymysql==1.0.2
    # via aiomysql
python-dateutil==2.8.2
    # via arrow
python-dotenv==0.20.0
    # via -r requirements.in
requests==2.28.1
    # via
    #   requests-oauthlib
    #   tweepy
requests-oauthlib==1.3.1
    # via tweepy
six==1.16.0
    # via
    #   astunparse
    #   python-dateutil
tomli==2.0.1
    # via black
tweepy[async]==4.10.0
    # via -r requirements.in
urllib3==1.26.10
    # via requests
uvloop==0.16.0
    # via -r requirements.in
wheel==0.37.1
    # via astunparse
yarl==1.7.2
    # via aiohttp
